Move SType definition into its own file'
