+++ src/core/types.h
@@ -13,6 +13,7 @@
 #include <limits>    // std::numeric_limits
 #include <string>    // std::string
 #include "python/python.h"
+#include "stype.h"
 using size_t = std::size_t;
 
 struct CString {
@@ -141,222 +142,6 @@ enum class LType : uint8_t {
 constexpr size_t DT_LTYPES_COUNT = static_cast<size_t>(LType::OBJECT) + 1;
 
 
-
-//==============================================================================
-
-/**
- * "Storage" type of a data column.
- *
- * These storage types are in 1-to-many correspondence with the logical types.
- * That is, a single logical type may have multiple storage types, but not the
- * other way around.
- *
- * SType::VOID
- *     "Fake" type, column with this stype is in an invalid state. This type may
- *     be used internally to indicate that a column is being constructed.
- *
- * -----------------------------------------------------------------------------
- *
- * SType::BOOL
- *     elem: int8_t (1 byte)
- *     NA:   -128
- *     A boolean with True = 1, False = 0. All other values are invalid.
- *
- * -----------------------------------------------------------------------------
- *
- * SType::INT8
- *     elem: int8_t (1 byte)
- *     NA:   -2**7 = -128
- *     An integer in the range -127 .. 127.
- *
- * SType::INT16
- *     elem: int16_t (2 bytes)
- *     NA:   -2**15 = -32768
- *     An integer in the range -32767 .. 32767.
- *
- * SType::INT32
- *     elem: int32_t (4 bytes)
- *     NA:   -2**31 = -2147483648
- *     An integer in the range -2147483647 .. 2147483647.
- *
- * SType::INT64
- *     elem: int64_t (8 bytes)
- *     NA:   -2**63 = -9223372036854775808
- *     An integer in the range -9223372036854775807 .. 9223372036854775807.
- *
- * -----------------------------------------------------------------------------
- *
- * SType::FLOAT32
- *     elem: float (4 bytes)
- *     NA:   0x7F8007A2
- *     Floating-point real number, corresponding to C's `float` (IEEE 754). We
- *     designate a specific NAN payload to mean the NA value; whereas all other
- *     numbers starting with 0x7F8 or 0xFF8 should be treated as actual NANs (or
- *     infinities).
- *
- * SType::FLOAT64
- *     elem: double (8 bytes)
- *     NA:   0x7FF00000000007A2
- *     Floating-point real number, corresponding to C's `double` (IEEE 754).
- *
- * SType::DEC16
- *     elem: int16_t (2 bytes)
- *     NA:   -2**15 = -32768
- *     The fixed-point real number (aka decimal); the scale variable in the meta
- *     indicates the number of digits after the decimal point. For example,
- *     number 7.11 can be stored as integer 711 with scale = 2.
- *     Note that this is different from IEEE 754 "decimal" format, since we
- *     include scale into the meta information of the column, rather than into
- *     each value. Thus, all values will have common scale, which greatly
- *     simplifies their use.
- *
- * SType::DEC32
- *     elem: int32_t (4 bytes)
- *     NA:   -2**31
- *     Fixed-point real number stored as a 32-bit integer.
- *
- * SType::DEC64
- *     elem: int64_t (8 bytes)
- *     NA:   -2**63
- *     Fixed-point real number stored as a 64-bit integer.
- *
- * -----------------------------------------------------------------------------
- *
- * SType::STR32
- *     elem: uint32_t (4 bytes) + unsigned char[]
- *     NA:   (1 << 31) mask
- *     Variable-width strings. The data consists of 2 buffers:
- *       1) string data. All non-NA strings are UTF-8 encoded and placed e
- *          nd-to-end.
- *       2) offsets. This is an array of uint32_t's representing offsets of
- *          each string in buffer 1). This array has nrows + 1 elements, with
- *          first element being 0, second element containing the end offset of
- *          the first string, ..., until the last element which stores the end
- *          offset of the last string. NA strings are encoded as the previous
- *          offset + NA mask (which is 1<<31).
- *     Thus, i-th string is NA if its highest bit is set, or otherwise it's a
- *     valid string whose starting offset is `start(i) = off(i-1) & ~(1<<31)`,
- *     ending offset is `end(i) = off(i)`, and `len(i) = end(i) - start(i)`.
- *
- *     For example, a column with 4 values `[NA, "hello", "", NA]` will be
- *     encoded as a string buffer of size 5 = (0 + 5 + 0 + 0) and offsets
- *     buffer containing 5 uint32_t's:
- *         strbuf  = [h e l l o]
- *         offsets = [0, 1<<31, 5, 5, 5|(1<<31)]
- *
- * SType::STR64
- *     elem: uint64_t (8 bytes) + unsigned char[]
- *     NA:   (1 << 63) mask
- *     Variable-width strings: same as SType::STR32 but use 64-bit
- *     offsets.
- *
- * SType::FSTR
- *     elem: unsigned char[] (n bytes)
- *     NA:   0xFF 0xFF ... 0xFF
- *     Fixed-width strings, similar to "CHAR(n)" in SQL. These strings have
- *     constant width `n` and are therefore stored as `char[n]` arrays. They are
- *     *not* null-terminated, however strings that are shorter than `n` in width
- *     will be \0-padded. The width `n` > 0 is given in the metadata. String
- *     data is encoded in UTF-8. The NA value is encoded as a sequence of 0xFF
- *     bytes, which is not a valid UTF-8 string.
- *
- * SType::CAT8
- *     elem: uint8_t (1 byte)
- *     NA:   255
- *     String column stored as a categorical variable (aka "factor" or "enum").
- *     This type is suitable for columns with low cardinality, i.e. having no
- *     more than 255 distinct string values.
- *     The data is stored in a buffer with 3 regions: first comes the string
- *     data, padded to a multiple of 8 bytes; then comes the array of int32
- *     offsets within the data buffer, these offsets describe the boundaries of
- *     individual strings within the data section (this array is also padded
- *     to a multiple of 8 bytes); finally comes the array of categorical
- *     indices. Meta information contains offsets of the second and the third
- *     sections. The layout of the first 2 sections is exactly the same as
- *     that of the SType::STR32 type.
- *
- * SType::CAT16
- *     elem: uint16_t (2 bytes)
- *     NA:   65535
- *     Strings stored as a categorical variable with no more than 65535 distinct
- *     levels. The layout is exactly the same as that of CAT8, only
- *     the last section uses 2 bytes per element instead of just 1 byte.
- *
- * SType::CAT32
- *     elem: uint32_t (4 bytes)
- *     NA:   2**32-1
- *     Strings stored as a categorical variable with no more than 2**32-1
- *     distinct levels. (The combined size of all categorical strings may not
- *     exceed 2**32 too). The layout is same as that of CAT8, only
- *     the last section uses 4 bytes per element instead of just 1 byte.
- *
- *
- * -----------------------------------------------------------------------------
- *
- * SType::DATE64
- *     elem: int64_t (8 bytes)
- *     NA:   -2**63
- *     Timestamp, stored as the number of microseconds since 0000-03-01. The
- *     allowed time range is ≈290,000 years around the epoch. The time is
- *     assumed to be in UTC, and does not allow specifying a time zone.
- *
- * SType::DATE32
- *     elem: int32_t (4 bytes)
- *     NA:   -2**31
- *     Date only: the number of days since 0000-03-01. The allowed time range
- *     is ≈245,000 years.
- *
- * SType::DATE16
- *     elem: int16_t (2 bytes)
- *     NA:   -2**15
- *     Year+month only: the number of months since 0000-03-01. The allowed time
- *     range is up to year 2730.
- *     This type is specifically designed for business applications. It allows
- *     adding/subtraction in monthly/yearly intervals (other datetime types do
- *     not allow that since months/years have uneven lengths).
- *
- * SType::TIME32
- *     elem: int32_t (4 bytes)
- *     NA:   -2**31
- *     Time only: the number of milliseconds since midnight. The allowed time
- *     range is ≈24 days.
- *
- *
- * -----------------------------------------------------------------------------
- *
- * SType::OBJ
- *     elem: PyObject*
- *     NA:   Py_None
- *
- */
-enum class SType : uint8_t {
-  VOID    = 0,
-  BOOL    = 1,
-  INT8    = 2,
-  INT16   = 3,
-  INT32   = 4,
-  INT64   = 5,
-  FLOAT32 = 6,
-  FLOAT64 = 7,
-  DEC16   = 8,
-  DEC32   = 9,
-  DEC64   = 10,
-  STR32   = 11,
-  STR64   = 12,
-  FSTR    = 13,
-  CAT8    = 14,
-  CAT16   = 15,
-  CAT32   = 16,
-  DATE64  = 17,
-  TIME32  = 18,
-  DATE32  = 19,
-  DATE16  = 20,
-  OBJ     = 21,
-
-  INVALID = 22,
-};
-
-constexpr size_t DT_STYPES_COUNT = static_cast<size_t>(SType::INVALID);
